requirements.txt

torch==2.2.0
torchvision==0.17.0
packaging
sentencepiece
peft==0.11.1
einops==0.8.0
fastapi==0.100.1
gradio==3.39.0
markdown2==2.4.10
numpy==1.24.2
openai==0.27.8
opencv_python==4.8.0.74
Pillow==9.4.0
pycocotools==2.0.6
ray==2.6.1
Requests==2.31.0
shortuuid==1.0.11
tqdm==4.64.1
transformers==4.31.0
uvicorn==0.23.2
scipy==1.11.2
bitsandbytes==0.41.1
deepspeed==0.14.3